Self-cleaning obsolete filesystems in FreeBSD plugin
