+++ src/freebsd_sysctl.c
@@ -29,12 +29,6 @@
 // --------------------------------------------------------------------------------------------------------------------
 // common definitions and variables
 
-#define KILO_FACTOR 1024
-#define MEGA_FACTOR 1048576     // 1024 * 1024
-#define GIGA_FACTOR 1073741824  // 1024 * 1024 * 1024
-
-#define MAX_INT_DIGITS 10 // maximum number of digits for int
-
 int system_pagesize = PAGE_SIZE;
 int number_of_cpus = 1;
 
@@ -2778,212 +2772,6 @@ int do_net_inet6_icmp6_stats(int update_every, usec_t dt) {
         }
     } else {
         error("DISABLED: net.inet6.icmp6.stats module");
-        return 1;
-    }
-
-    return 0;
-}
-
-// --------------------------------------------------------------------------------------------------------------------
-// getmntinfo
-
-int do_getmntinfo(int update_every, usec_t dt) {
-    (void)dt;
-
-#define DELAULT_EXLUDED_PATHS "/proc/*"
-// taken from gnulib/mountlist.c and shortened to FreeBSD related fstypes
-#define DEFAULT_EXCLUDED_FILESYSTEMS "autofs procfs subfs devfs none"
-#define CONFIG_SECTION_GETMNTINFO "plugin:freebsd:getmntinfo"
-
-    static int do_space = -1, do_inodes = -1;
-
-    if (unlikely(do_space == -1)) {
-        do_space  = config_get_boolean_ondemand(CONFIG_SECTION_GETMNTINFO, "space usage for all disks",  CONFIG_BOOLEAN_AUTO);
-        do_inodes = config_get_boolean_ondemand(CONFIG_SECTION_GETMNTINFO, "inodes usage for all disks", CONFIG_BOOLEAN_AUTO);
-    }
-
-    if (likely(do_space || do_inodes)) {
-        struct statfs *mntbuf;
-        int mntsize;
-
-        // there is no mount info in sysctl MIBs
-        if (unlikely(!(mntsize = getmntinfo(&mntbuf, MNT_NOWAIT)))) {
-            error("FREEBSD: getmntinfo() failed");
-            do_space = 0;
-            error("DISABLED: disk_space.* charts");
-            do_inodes = 0;
-            error("DISABLED: disk_inodes.* charts");
-            error("DISABLED: getmntinfo module");
-            return 1;
-        } else {
-            // Data to be stored in DICTIONARY mount_points.
-            // This DICTIONARY is used to lookup the settings of the mount point on each iteration.
-            struct mount_point_metadata {
-                int do_space;
-                int do_inodes;
-
-                size_t collected; // the number of times this has been collected
-
-                // charts and dimensions
-
-                RRDSET *st_space;
-                RRDDIM *rd_space_used;
-                RRDDIM *rd_space_avail;
-                RRDDIM *rd_space_reserved;
-
-                RRDSET *st_inodes;
-                RRDDIM *rd_inodes_used;
-                RRDDIM *rd_inodes_avail;
-            };
-            static DICTIONARY *mount_points = NULL;
-            static SIMPLE_PATTERN *excluded_mountpoints = NULL;
-            static SIMPLE_PATTERN *excluded_filesystems = NULL;
-            int i;
-
-            if(unlikely(!mount_points)) {
-
-                excluded_mountpoints = simple_pattern_create(
-                        config_get(CONFIG_SECTION_GETMNTINFO, "exclude space metrics on paths",
-                                   DELAULT_EXLUDED_PATHS),
-                        SIMPLE_PATTERN_EXACT
-                );
-
-                excluded_filesystems = simple_pattern_create(
-                        config_get(CONFIG_SECTION_GETMNTINFO, "exclude space metrics on filesystems",
-                                   DEFAULT_EXCLUDED_FILESYSTEMS),
-                        SIMPLE_PATTERN_EXACT
-                );
-
-                mount_points = dictionary_create(DICTIONARY_FLAG_SINGLE_THREADED);
-            }
-
-            for (i = 0; i < mntsize; i++) {
-
-                char title[4096 + 1];
-                int def_space, def_inodes, iter_space, iter_inodes;
-
-                struct mount_point_metadata *m = dictionary_get(mount_points, mntbuf[i].f_mntonname);
-                if(unlikely(!m)) {
-                    char var_name[4096 + 1];
-                    snprintfz(var_name, 4096, "%s:%s", CONFIG_SECTION_GETMNTINFO, mntbuf[i].f_mntonname);
-
-                    def_space = do_space;
-                    def_inodes = do_space;
-
-                    if(unlikely(simple_pattern_matches(excluded_mountpoints, mntbuf[i].f_mntonname))) {
-                        def_space = CONFIG_BOOLEAN_NO;
-                        def_inodes = CONFIG_BOOLEAN_NO;
-                    }
-
-                    if(unlikely(simple_pattern_matches(excluded_filesystems, mntbuf[i].f_fstypename))) {
-                        def_space = CONFIG_BOOLEAN_NO;
-                        def_inodes = CONFIG_BOOLEAN_NO;
-                    }
-
-                    iter_space  = config_get_boolean_ondemand(var_name, "space usage",  def_space);
-                    iter_inodes = config_get_boolean_ondemand(var_name, "inodes usage", def_inodes);
-
-                    struct mount_point_metadata mp = {
-                            .do_space = iter_space,
-                            .do_inodes = iter_inodes,
-
-                            .collected = 0,
-
-                            .st_space = NULL,
-                            .rd_space_avail = NULL,
-                            .rd_space_used = NULL,
-                            .rd_space_reserved = NULL,
-
-                            .st_inodes = NULL,
-                            .rd_inodes_avail = NULL,
-                            .rd_inodes_used = NULL,
-                    };
-
-                    m = dictionary_set(mount_points, mntbuf[i].f_mntonname, &mp, sizeof(struct mount_point_metadata));
-                }
-
-                if(unlikely(m->do_space == CONFIG_BOOLEAN_NO && m->do_inodes == CONFIG_BOOLEAN_NO))
-                    continue;
-
-                if(unlikely(mntbuf[i].f_flags & MNT_RDONLY && !m->collected))
-                    continue;
-
-                // --------------------------------------------------------------------------
-
-                int rendered = 0;
-
-                if (m->do_space == CONFIG_BOOLEAN_YES || (m->do_space == CONFIG_BOOLEAN_AUTO && (mntbuf[i].f_blocks > 2))) {
-                    if (unlikely(!m->st_space)) {
-                        snprintfz(title, 4096, "Disk Space Usage for %s [%s]",
-                                  mntbuf[i].f_mntonname, mntbuf[i].f_mntfromname);
-                        m->st_space = rrdset_create_localhost("disk_space",
-                                                              mntbuf[i].f_mntonname,
-                                                              NULL,
-                                                              mntbuf[i].f_mntonname,
-                                                              "disk.space",
-                                                              title,
-                                                              "GB",
-                                                              2023,
-                                                              update_every,
-                                                              RRDSET_TYPE_STACKED
-                        );
-
-                        m->rd_space_avail    = rrddim_add(m->st_space, "avail", NULL,
-                                                          mntbuf[i].f_bsize, GIGA_FACTOR, RRD_ALGORITHM_ABSOLUTE);
-                        m->rd_space_used     = rrddim_add(m->st_space, "used", NULL,
-                                                          mntbuf[i].f_bsize, GIGA_FACTOR, RRD_ALGORITHM_ABSOLUTE);
-                        m->rd_space_reserved = rrddim_add(m->st_space, "reserved_for_root", "reserved for root",
-                                                          mntbuf[i].f_bsize, GIGA_FACTOR, RRD_ALGORITHM_ABSOLUTE);
-                    } else
-                        rrdset_next(m->st_space);
-
-                    rrddim_set_by_pointer(m->st_space, m->rd_space_avail,    (collected_number) mntbuf[i].f_bavail);
-                    rrddim_set_by_pointer(m->st_space, m->rd_space_used,     (collected_number) (mntbuf[i].f_blocks -
-                                                                                                 mntbuf[i].f_bfree));
-                    rrddim_set_by_pointer(m->st_space, m->rd_space_reserved, (collected_number) (mntbuf[i].f_bfree -
-                                                                                                 mntbuf[i].f_bavail));
-                    rrdset_done(m->st_space);
-
-                    rendered++;
-                }
-
-                // --------------------------------------------------------------------------
-
-                if (m->do_inodes == CONFIG_BOOLEAN_YES || (m->do_inodes == CONFIG_BOOLEAN_AUTO && (mntbuf[i].f_files > 1))) {
-                    if (unlikely(!m->st_inodes)) {
-                        snprintfz(title, 4096, "Disk Files (inodes) Usage for %s [%s]",
-                                  mntbuf[i].f_mntonname, mntbuf[i].f_mntfromname);
-                        m->st_inodes = rrdset_create_localhost("disk_inodes",
-                                                               mntbuf[i].f_mntonname,
-                                                               NULL,
-                                                               mntbuf[i].f_mntonname,
-                                                               "disk.inodes",
-                                                               title,
-                                                               "Inodes",
-                                                               2024,
-                                                               update_every,
-                                                               RRDSET_TYPE_STACKED
-                        );
-
-                        m->rd_inodes_avail = rrddim_add(m->st_inodes, "avail", NULL, 1, 1, RRD_ALGORITHM_ABSOLUTE);
-                        m->rd_inodes_used  = rrddim_add(m->st_inodes, "used",  NULL, 1, 1, RRD_ALGORITHM_ABSOLUTE);
-                    } else
-                        rrdset_next(m->st_inodes);
-
-                    rrddim_set_by_pointer(m->st_inodes, m->rd_inodes_avail, (collected_number) mntbuf[i].f_ffree);
-                    rrddim_set_by_pointer(m->st_inodes, m->rd_inodes_used,  (collected_number) (mntbuf[i].f_files -
-                                                                                                mntbuf[i].f_ffree));
-                    rrdset_done(m->st_inodes);
-
-                    rendered++;
-                }
-
-                if(likely(rendered))
-                    m->collected++;
-            }
-        }
-    } else {
-        error("DISABLED: getmntinfo module");
         return 1;
     }
